Implemented some use cases of task entity. Organized the packaging of use cases.
