--- src/main/java/task_use_cases/CreateTask.java
@@ -1,4 +1,0 @@
-package task_use_cases;
-
-public class CreateTask {
-}
