+++ src/main/java/use_cases/tasks/DeleteCategoryPresenter.java
@@ -1,4 +1,4 @@
-package task_use_cases;
+package use_cases.tasks;
 
 public class DeleteCategoryPresenter {
 }
