+++ src/main/java/use_cases/tasks/DeleteTaskPresenter.java
@@ -1,4 +1,4 @@
-package task_use_cases;
+package use_cases.tasks;
 
 public class DeleteTaskPresenter {
 }
